Combine importxml.cpp and importmusicxml.cpp
